+++ lib/bill/bill/sat/solver/abc/abc_global.h
@@ -76,6 +76,7 @@
 #include <assert.h>
 #include <stdio.h>
 #include <string.h>
+#include <cstdint>
 
 // catch memory leaks in Visual Studio
 #ifdef WIN32
@@ -130,7 +131,7 @@ typedef unsigned long ABC_PTRUINT_T;
 #elif     defined(NT64)
 typedef unsigned long long ABC_PTRUINT_T;
 #elif     defined(NT) || defined(LIN) || defined(WIN32)
-typedef unsigned int ABC_PTRUINT_T;
+typedef uintptr_t ABC_PTRUINT_T;
 #else
    #error unknown platform
 #endif /* defined(PLATFORM) */
@@ -147,7 +148,7 @@ typedef long ABC_PTRINT_T;
 #elif     defined(NT64)
 typedef long long ABC_PTRINT_T;
 #elif     defined(NT) || defined(LIN) || defined(WIN32)
-typedef int ABC_PTRINT_T;
+typedef intptr_t ABC_PTRINT_T;
 #else
    #error unknown platform
 #endif /* defined(PLATFORM) */
